+++ django_stripe/admin/__init__.py
@@ -1,4 +1,4 @@
-from django_stripe.admin.core.customers import StripeCustomerAdmin
-from django_stripe.admin.products.products import StripeProductAdmin
+from django_stripe.admin.core import StripeCustomerAdmin
+from django_stripe.admin.products import StripeProductAdmin, StripeCouponAdmin
 
 __all__ = ("StripeCustomerAdmin", "StripeProductAdmin")