+++ django_stripe/admin/products/products.py
@@ -7,7 +7,7 @@ from django_stripe.models import StripeProduct, StripePrice
 from django_stripe.actions import StripeProductAction
 
 
-class StripeSubscriptionInlineAdmin(admin.StackedInline):
+class StripePriceInlineAdmin(admin.StackedInline):
     model = StripePrice
     extra = 0
 
@@ -16,5 +16,5 @@ class StripeSubscriptionInlineAdmin(admin.StackedInline):
 class StripeProductAdmin(AbstractStripeModelAdmin):
     list_display = ("stripe_id", "name", "description")
     search_fields = ("stripe_id", "name", "description")
-    inlines = [StripeSubscriptionInlineAdmin]
+    inlines = [StripePriceInlineAdmin]
     stripe_model_action = StripeProductAction()